+++ docs/whats-new.md
@@ -1,5 +1,8 @@
 ## UNRELEASED
 
+### Feat
+- moved from pickle to netcdf and improved statistics in [#14](https://github.com/Deltares-research/kenmerkendewaarden/pull/14)
+
 
 ## 0.1.0 (2024-03-11)
 This is the set of kenmerkende waarden kust scripts and functions as transfered from hatyan and how they were applied in the kwk-2022 project.